pyproject.toml
[tool.poetry]
name = "uranium_quantum"
version = "0.1.4"
description = "Support Python libraries for Uranium quantum computing platform."
authors = ["Radu Marginean <radu.marg@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://www.transilvania-quantum.org/uranium"
repository = "https://github.com/radumarg/uranium_quantum"
keywords = ["uranium", "quantum", "computing"]

include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.1.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
